pyproject.toml
[project]
name = "financecoding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = []

[dependency-groups]
data = [
    "beautifulsoup4>=4.13.4",
    "feedparser>=6.0.11",
    "lxml-html-clean>=0.4.2",
    "newspaper3k>=0.2.8",
    "playwright>=1.53.0",
    "requests>=2.32.4",
]
dev = [
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.1",
]
